back/requirements.txt
alembic==1.7.5
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
fsspec==2021.11.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
mysqlclient==2.1.0
numpy==1.22.0
pandas==1.3.5
pathlib==1.0.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
scipy==1.7.3
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.7.0
